refactor: extract cleanupTestDirectories to tests/test-utils.ts
- Create tests/test-utils.ts with shared cleanupTestDirectories function
- Add logPrefix parameter for customizable log messages
- Update global-setup.ts and global-teardown.ts to import from test-utils

Refs #1093
